dt-overlay-at91: check the .itb exists before installing it

+++ recipes-bsp/dt-overlay-at91/dt-overlay-at91_git.bb
@@ -68,7 +68,9 @@ addtask deploy after do_install
 
 do_deploy () {
     #echo "Copying ${fit_image_basename}.itb and source file to ${DEPLOYDIR}..."
-    install ${AT91BOOTSTRAP_MACHINE}.itb ${DEPLOYDIR}/
-    install ${AT91BOOTSTRAP_MACHINE}.its ${DEPLOYDIR}/
+    if [ -e ${AT91BOOTSTRAP_MACHINE}.itb ]; then
+        install ${AT91BOOTSTRAP_MACHINE}.itb ${DEPLOYDIR}/
+        install ${AT91BOOTSTRAP_MACHINE}.its ${DEPLOYDIR}/
+    fi;
 }
